requirements.txt
uvicorn==0.29.0
fastapi==0.110.2
fastapi-csrf-protect==0.2.1
motor==2.5.1
PyJWT==2.3.0
passlib==1.7.4
python-decouple==3.3
dnspython==2.1.0
gunicorn==20.1.0
mysqlclient==2.2.4
sqlalchemy==1.4.40
python-dateutil==2.9.0.post0
python-jose==3.3.0
python-multipart==0.0.9
bcrypt==4.1.2
pydantic==1.10.15
pytz==2024.1
python-dotenv
azure-storage-blob
fastapi-csrf-protect
azure-storage-file-share
